AddRegisterGenerator
